don't leak subtle cursor mode from mc-focus
So, subtle cursor mode is currently global.  It kind of needs to be local, but
that is also it's own bag of worms

This is the simplest case.  If MC Focus buffer dies and suble cursor wasn't
active when mc-focus last started, we turn off any subtle cursor mode

I looked at toggling the mode on window buffer changes, but the number of edge
cases escaped my first pass and it was becoming silly.

If subtle cursor was buffer local, this wouldn't be an issue anyway.