Obscurring suppport
Immediately after writing this, I considerred adding support to evaporate part
of the buffer text, shrinking it rather than obscurring it.

In a lot of casese, without re-flow and re-indentation etc, The results will not
be amazing.

When working with balanced s-expressions, it might be preferable to work more
like org edit, which would give us a proper mode.

Combining babel blocks, these workflows are not particularly time consuming.
Would it be worth it to do what babel does, using an indirect buffer in another
mode?

Not always.  Only balanced expressions will work as intended when activating a
mode.  moc-focus is strong in cases where you want to just display the buffer
text, regardless of how it is balanced or not.

Well, enough thinking about it.  Enjoy your new feature.