hide fringes.  set fringes to have no background color.
This could potentially be done only in the echo buffer, but I need to be able to
find that buffer.

In general, tracking frames is kind of heavy on an application, and I'd like to
keep MC to one instance for now.  It doesn't support not taking over the frame.
It is very focused, and adding lots of frame tracking will not be good until
it's time to attack that as its own can of worms.