Filter out the region overlay
We do not wish to have a permanent region highlight in the results =D

CL loop >>>>>>> Emacs loops.

We're adopting CL.  I'm tired of it.  Every single time I need to restructure
even the tiniest list, the built-in Elisp loop macros act like a dying mouse in
a mouse trap, too feeble to even reach the cheese in order to meet their fate.
The Elisp loop macros are comparatively useless.  `while-let' is okay.  That's
the only one who can stay.  `mapc' is okay.  `dolist', `dotimes' etc are just
elementary garbage I would not wish upon anyone in the modern era.

That is all.  CL is my friend now.  No longer friends with... I was never
friends with.. any of them.