Re-implement extraction for future-proofing
The excerpt to be displayed has a lot of concerns to address.  To avoid doing
math and stitching overlays, a buffer based solution was used.

Wherever Emacs is smart about columns, line endings, whitespace etc, so will be
MoC.  This is much more maintainable.

By editing in the buffer where overlays have been propagated, the overlay
stitching problem was completely avoided and C is probably faster at this and
happy not to be managing the garbage anyway.

This is the way of the future.