continuation strategy implementation: truncation
The idea of window text area has to change when we know we will truncate lines.
This affects how large we want to make the text as well.  We need to consider
post-truncation area.

Because truncate-lines uses the window's column width, we need to switch back to
margins when the continuation strategy is truncate

🚧 This commit does not include an implemenation of visual lines or single
lines.