Check for visual wrap mode before attempting to activate
The following case can occur:  A user records an expression with one mode
active.  On a later Emacs version or another user's Emacs, they have a diffent
version of `visual-wrap' that contains or does not contain the same functions.
These two modes do the equivalent thing.  We do support the upgrade path, from
adaptive-wrap-prefix mode to visual-wrap-prefix-mode, but not visa versa.